refactoring

+++ ansible/roles/kubernetes/files/genbootcmd.sh
@@ -11,16 +11,25 @@ rpilogit () {
 
 # run as root
 if [[ root != "$(whoami)" ]]; then
-  echo "Error: requires root";
+  rpilogit "Error: genbootcmd.sh requires root";
   exit 1;
 fi
 
+# check for /boot/cmdline.txt
+if [ ! -f /boot/cmdline.txt ]; then
+	rpilogit "ERROR: missing cmdline.txt";
+	exit 1;
+fi
+
+
 rpilogit "starting genbootcmd.sh"
+
 
 DISKMAIN=$(blkid -o export /dev/mmcblk0p2 | grep "PARTUUID")
 
 BOOTOPS="dwc_otg.lpm_enable=0 console=serial0,115200 console=tty1 root=${DISKMAIN} rootfstype=ext4 cgroup_enable=memory cgroup_memory=1 elevator=deadline fsck.repair=yes rootwait"
 
 echo $BOOTOPS > /boot/cmdline.txt;
+
 
 rpilogit "finished genbootcmd.sh"
